make event marginally more useful
