+++ src/main/java/io/github/fablabsmc/fablabs/api/fluidvolume/v1/FluidHelper.java
@@ -3,9 +3,9 @@ package io.github.fablabsmc.fablabs.api.fluidvolume.v1;
 import java.util.ArrayList;
 import java.util.List;
 
-import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.EntityFluidContainer;
-import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.ItemFluidContainer;
-import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.SidedFluidContainer;
+import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.EntityFluidContainerProvider;
+import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.ItemFluidContainerProvider;
+import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.BlockFluidContainerProvider;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.ImmutableFluidVolume;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.MultiFluidContainer;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.api.FluidContainer;
@@ -21,7 +21,7 @@ import net.minecraft.util.math.Box;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
 
-public final class FluidView {
+public final class FluidHelper {
 	/**
 	 * attempt to get a fluid container from the world from a specific block (or entities) for a given face.
 	 *
@@ -36,20 +36,20 @@ public final class FluidView {
 
 		FluidContainer container = ImmutableFluidVolume.EMPTY;
 
-		if (block instanceof SidedFluidContainer) {
-			container = ((SidedFluidContainer) block).getContainer(world, state, pos, side);
+		if (block instanceof BlockFluidContainerProvider) {
+			container = ((BlockFluidContainerProvider) block).getContainer(world, state, pos, side);
 		} else {
-			SidedFluidContainer sidedContainer = WorldFluidCallback.EVENT.invoker().getContainer(world, pos);
+			FluidContainer worldContainer = WorldFluidCallback.EVENT.invoker().getContainer(world, pos);
 
-			if (sidedContainer != null) {
-				container = sidedContainer.getContainer(world, state, pos, side);
+			if (worldContainer != null) {
+				container = new MultiFluidContainer(container, worldContainer);
 			}
 		}
 
 		List<FluidContainer> containers = new ArrayList<>();
 
-		for (Entity entity : world.getEntities((Entity) null, new Box(pos), e -> e instanceof EntityFluidContainer)) {
-			containers.add(((EntityFluidContainer) entity).getContainer(side));
+		for (Entity entity : world.getEntities((Entity) null, new Box(pos), e -> e instanceof EntityFluidContainerProvider)) {
+			containers.add(((EntityFluidContainerProvider) entity).getContainer(side));
 		}
 
 		if (!containers.isEmpty()) {
@@ -70,8 +70,8 @@ public final class FluidView {
 	public static FluidContainer getContainer(ItemStack stack) {
 		Item item = stack.getItem();
 
-		if (item instanceof ItemFluidContainer) {
-			FluidContainer container = ((ItemFluidContainer) item).getVolume(stack);
+		if (item instanceof ItemFluidContainerProvider) {
+			FluidContainer container = ((ItemFluidContainerProvider) item).getVolume(stack);
 
 			if (container != null) {
 				return container;
