+++ src/main/java/io/github/fablabsmc/fablabs/api/fluidvolume/v1/containers/EntityFluidContainerProvider.java
@@ -7,7 +7,7 @@ import net.minecraft.util.math.Direction;
 /**
  * an entity that carries fluids.
  */
-public interface EntityFluidContainer {
+public interface EntityFluidContainerProvider {
 	/**
 	 * @return the fluid container for this entity
 	 */
