+++ src/main/java/io/github/fablabsmc/fablabs/api/fluidvolume/v1/mixin/CauldronBlockMixin.java
@@ -1,6 +1,6 @@
 package io.github.fablabsmc.fablabs.api.fluidvolume.v1.mixin;
 
-import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.SidedFluidContainer;
+import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.BlockFluidContainerProvider;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.api.FluidVolume;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.vanilla.CauldronFluidVolume;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,7 +13,7 @@ import net.minecraft.world.World;
 
 @Mixin(CauldronBlock.class)
 // cauldrons can hold fluids (well, just water), in fact they're the only vanilla block that stores them as far as I'm aware
-public class CauldronBlockMixin implements SidedFluidContainer {
+public class CauldronBlockMixin implements BlockFluidContainerProvider {
 	@Override
 	public FluidVolume getContainer(World world, BlockState state, BlockPos pos, Direction side) {
 		return new CauldronFluidVolume(pos, world, state);
