+++ src/main/java/io/github/fablabsmc/fablabs/api/fluidvolume/v1/mixin/GlassBottleItemMixin.java
@@ -1,6 +1,6 @@
 package io.github.fablabsmc.fablabs.api.fluidvolume.v1.mixin;
 
-import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.ItemFluidContainer;
+import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.ItemFluidContainerProvider;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.ImmutableFluidVolume;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.api.FluidContainer;
 import org.spongepowered.asm.mixin.Mixin;
@@ -10,7 +10,7 @@ import net.minecraft.item.ItemStack;
 
 @Mixin(GlassBottleItem.class)
 // glass bottles do not hold fluid, but they can once they are filled, however they technically become a different item
-public class GlassBottleItemMixin implements ItemFluidContainer {
+public class GlassBottleItemMixin implements ItemFluidContainerProvider {
 	@Override
 	public FluidContainer getVolume(ItemStack stack) {
 		return ImmutableFluidVolume.EMPTY;
