+++ src/main/java/io/github/fablabsmc/fablabs/api/fluidvolume/v1/mixin/PotionItemMixin.java
@@ -1,6 +1,6 @@
 package io.github.fablabsmc.fablabs.api.fluidvolume.v1.mixin;
 
-import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.ItemFluidContainer;
+import io.github.fablabsmc.fablabs.api.fluidvolume.v1.containers.ItemFluidContainerProvider;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.api.FluidContainer;
 import io.github.fablabsmc.fablabs.api.fluidvolume.v1.volume.vanilla.PotionFluidVolume;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,7 +9,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.PotionItem;
 
 @Mixin(PotionItem.class) // potions hold fluid
-public class PotionItemMixin implements ItemFluidContainer {
+public class PotionItemMixin implements ItemFluidContainerProvider {
 	@Override
 	public FluidContainer getVolume(ItemStack stack) {
 		return new PotionFluidVolume(stack);
